requirements.txt
numpy
pandas 
scikit-learn
tensorflow
keras-tcn
mediapipe
opencv-python
joblib
google-cloud-sdk